# perf improvements to CalculatePOC