Old code:
```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/sharding
BenchmarkCalculatePOC-12             100          11704937 ns/op        14080345 B/op     121660 allocs/op
PASS
```

New code:
```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/sharding
BenchmarkCalculatePOC-12             100          10932186 ns/op        13344667 B/op     115783 allocs/op
PASS
```

The change is modest because most of the time in the function is spent merklizing the tree (calling `DeriveSha`). Commenting out the call and running just the construction of the collation body with the salt makes the change more drastic.

Old code:
```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/sharding
BenchmarkCalculatePOC-12          100000             11543 ns/op           40064 B/op        309 allocs/op
PASS
```

New code:
```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/sharding
BenchmarkCalculatePOC-12         1000000              1727 ns/op            6528 B/op          1 allocs/op
PASS
```

Notice that there's only 1 alloc with the new code! That's because all the bytes are allocated at the start and we avoid unnecessarily copying data around.